Update to tape-dispenser fork for separate publishing
